--- qiskit_ibm_runtime/executor/routines/sampler_v2/post_processors/__init__.py
@@ -1,16 +1,0 @@
-# This code is part of Qiskit.
-#
-# (C) Copyright IBM 2026.
-#
-# This code is licensed under the Apache License, Version 2.0. You may
-# obtain a copy of this license in the LICENSE.txt file in the root directory
-# of this source tree or at http://www.apache.org/licenses/LICENSE-2.0.
-#
-# Any modifications or derivative works of this code must retain this
-# copyright notice, and modified files need to carry a notice indicating
-# that they have been altered from the originals.
-
-"""Post-processors for executor-based SamplerV2."""
-
-from .utils import SAMPLER_POST_PROCESSORS
-from .v0_1.post_processor import sampler_v2_post_processor_v0_1
