+++ qiskit_ibm_runtime/sampler_v2/post_processors/__init__.py
@@ -10,4 +10,4 @@
 # copyright notice, and modified files need to carry a notice indicating
 # that they have been altered from the originals.
 
-"""Tests for executor routines."""
+"""V0.1 post processor for executor-based SamplerV2"""
