+++ test/integration/test_executor.py
@@ -23,7 +23,7 @@ from samplomatic.transpiler import generate_boxing_pass_manager
 
 from qiskit_ibm_runtime import Executor, QuantumProgram
 from qiskit_ibm_runtime.quantum_program import QuantumProgramResult
-from ...ibm_test_case import IBMIntegrationTestCase
+from ..ibm_test_case import IBMIntegrationTestCase
 
 
 class TestExecutor(IBMIntegrationTestCase):
