--- test/unit/executor/routines/sampler_v2/post_processors/test_sampler_post_processors.py
@@ -1,20 +1,0 @@
-# This code is part of Qiskit.
-#
-# (C) Copyright IBM 2026.
-#
-# This code is licensed under the Apache License, Version 2.0. You may
-# obtain a copy of this license in the LICENSE.txt file in the root directory
-# of this source tree or at http://www.apache.org/licenses/LICENSE-2.0.
-#
-# Any modifications or derivative works of this code must retain this
-# copyright notice, and modified files need to carry a notice indicating
-# that they have been altered from the originals.
-
-"""Tests for SamplerV2 post-processors."""
-
-from qiskit_ibm_runtime.executor.routines.sampler_v2.post_processors import SAMPLER_POST_PROCESSORS
-
-
-def test_available_post_processors():
-    """Test the available post-processors."""
-    assert "v0.1" in SAMPLER_POST_PROCESSORS
