+++ test/unit/sampler_v2/post_processors/test_flatten_twirling_axes.py
@@ -18,7 +18,7 @@ import unittest
 
 import numpy as np
 
-from qiskit_ibm_runtime.executor.routines.sampler_v2.post_processors.v0_1.flatten_twirling_axes import (  # noqa: E501
+from qiskit_ibm_runtime.sampler_v2.post_processors.utils import (
     flatten_twirling_axes,
 )
